+++ hedera-mirror-importer/src/test/java/com/hedera/mirror/importer/parser/record/entity/redis/RedisEntityListenerTest.java
@@ -20,49 +20,127 @@ package com.hedera.mirror.importer.parser.record.entity.redis;
  * ‍
  */
 
-import org.junit.jupiter.api.TestInstance;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.data.redis.core.RedisCallback;
-import org.springframework.data.redis.core.RedisOperations;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.serializer.RedisSerializer;
-import reactor.core.publisher.Flux;
-import reactor.core.publisher.UnicastProcessor;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.timeout;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
+import io.micrometer.core.instrument.simple.SimpleMeterRegistry;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.atomic.AtomicInteger;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.Mockito;
+import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.data.redis.core.RedisOperations;
+import org.springframework.data.redis.core.SessionCallback;
+
+import com.hedera.mirror.importer.MirrorProperties;
 import com.hedera.mirror.importer.domain.StreamMessage;
 import com.hedera.mirror.importer.domain.TopicMessage;
-import com.hedera.mirror.importer.parser.record.entity.BatchEntityListenerTest;
+import com.hedera.mirror.importer.parser.record.entity.EntityBatchCleanupEvent;
+import com.hedera.mirror.importer.parser.record.entity.EntityBatchSaveEvent;
 
-@SpringBootTest(properties = {
-        "hedera.mirror.importer.parser.record.entity.redis.enabled=true"})
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
-class RedisEntityListenerTest extends BatchEntityListenerTest {
+@ExtendWith(MockitoExtension.class)
+class RedisEntityListenerTest {
 
-    private final RedisOperations<String, StreamMessage> redisOperations;
+    @Mock
+    private RedisOperations<String, StreamMessage> redisOperations;
 
-    @Autowired
-    public RedisEntityListenerTest(RedisEntityListener entityListener, RedisProperties properties,
-                                   RedisOperations<String, StreamMessage> redisOperations) {
-        super(entityListener, properties);
-        this.redisOperations = redisOperations;
+    private RedisEntityListener entityListener;
+
+    private long consensusTimestamp = 1;
+
+    @BeforeEach
+    void setup() {
+        entityListener = new RedisEntityListener(new MirrorProperties(),
+                new RedisProperties(), redisOperations, new SimpleMeterRegistry());
+        entityListener.init();
     }
 
-    @Override
-    protected Flux<TopicMessage> subscribe(long topicNum) {
-        UnicastProcessor<TopicMessage> processor = UnicastProcessor.create();
-        RedisSerializer stringSerializer = ((RedisTemplate<String, ?>) redisOperations).getStringSerializer();
-        RedisSerializer<TopicMessage> serializer = (RedisSerializer<TopicMessage>) redisOperations.getValueSerializer();
+    @Test
+    void onSlowPublish() {
+        CountDownLatch latch = new CountDownLatch(1);
+        AtomicInteger saveCount = new AtomicInteger(0);
 
-        RedisCallback<TopicMessage> redisCallback = connection -> {
-            byte[] channel = stringSerializer.serialize("topic.0.0." + topicNum);
-            connection.subscribe((message, pattern) -> {
-                processor.onNext(serializer.deserialize(message.getBody()));
-            }, channel);
+        // given
+        List<TopicMessage> messages = new ArrayList<>();
+        for (int i = 0; i < RedisEntityListener.TASK_QUEUE_SIZE + 2; i++) {
+            messages.add(topicMessage());
+        }
+
+        // when
+        when(redisOperations.executePipelined(any(SessionCallback.class))).then((callback) -> {
+            latch.await();
             return null;
-        };
+        });
 
-        redisOperations.execute(redisCallback);
-        return processor;
+        // drive entityListener to handle topic messages in a different thread cause it will block
+        new Thread(() -> {
+            try {
+                for (TopicMessage message : messages) {
+                    submitAndSave(message);
+                    entityListener.onCleanup(new EntityBatchCleanupEvent(this));
+                    saveCount.getAndIncrement();
+                }
+            } catch (InterruptedException e) {
+                // ignore
+            }
+        }).start();
+
+        //then
+        //Thread is blocked because queue is full, and publisher is blocked on first message.
+        verify(redisOperations, timeout(500).times(1))
+                .executePipelined(any(SessionCallback.class));
+        assertThat(saveCount.get()).isEqualTo(RedisEntityListener.TASK_QUEUE_SIZE + 1);
+
+        latch.countDown();
+        //All messages should be queued and published
+        verify(redisOperations, timeout(500).times(RedisEntityListener.TASK_QUEUE_SIZE + 2))
+                .executePipelined(any(SessionCallback.class));
+        assertThat(saveCount.get()).isEqualTo(RedisEntityListener.TASK_QUEUE_SIZE + 2);
+    }
+
+    @Test
+    void onDuplicateTopicMessages() throws InterruptedException {
+        TopicMessage topicMessage1 = topicMessage();
+        TopicMessage topicMessage2 = topicMessage();
+        TopicMessage topicMessage3 = topicMessage();
+
+        //submitAndSave two messages, verify publish logic called twice
+        submitAndSave(topicMessage1);
+        submitAndSave(topicMessage2);
+        verify(redisOperations, timeout(500).times(2))
+                .executePipelined(any(SessionCallback.class));
+
+        //submitAndSave two duplicate messages, verify publish was not attempted
+        Mockito.reset(redisOperations);
+        submitAndSave(topicMessage1);
+        submitAndSave(topicMessage2);
+        verify(redisOperations, timeout(500).times(0))
+                .executePipelined(any(SessionCallback.class));
+
+        //submitAndSave third new unique message, verify publish called once.
+        Mockito.reset(redisOperations);
+        submitAndSave(topicMessage3);
+        entityListener.onCleanup(new EntityBatchCleanupEvent(this));
+        verify(redisOperations, timeout(500).times(1))
+                .executePipelined(any(SessionCallback.class));
+    }
+
+    private TopicMessage topicMessage() {
+        TopicMessage topicMessage = new TopicMessage();
+        topicMessage.setConsensusTimestamp(consensusTimestamp++);
+        return topicMessage;
+    }
+
+    private void submitAndSave(TopicMessage topicMessage) throws InterruptedException {
+        entityListener.onTopicMessage(topicMessage);
+        entityListener.onSave(new EntityBatchSaveEvent(this));
     }
 }
